+++ src/pcc_qa/pcc/CephRbd.py
@@ -89,11 +89,12 @@ class CephRbd(PccBase):
 
         response = pcc.get_ceph_rbds(conn)
         for data in get_response_data(response):
-            response=self.delete_ceph_rbd_by_id(id=data['id'])
-            status=self.wait_until_rbd_deleted(id=data['id'])
-            if status!="OK":
-                print("{} deletion failed".format(data['name']))
-                return "Error"
+            if data["managed"] == True:
+                response=self.delete_ceph_rbd_by_id(id=data['id'])
+                status=self.wait_until_rbd_deleted(id=data['id'])
+                if status!="OK":
+                    print("{} deletion failed".format(data['name']))
+                    return "Error"
         return "OK"
 
     ###########################################################################
